Added consistency test between C and Fortran versions of spherical surface fitting codes.

+++ extern/dierckx/src/fpgivs.c
@@ -8,16 +8,16 @@ extern "C" {
 void fpgivs(float piv, float *ww, float *cos_, float *sin_)
 {
     float dd, store, t;
-    store = fabs(piv);
+    store = fabsf(piv);
     if (store < *ww)
     {
         t = piv / *ww;
-        dd = *ww * sqrt(1.0 + t * t);
+        dd = *ww * sqrtf(1.0 + t * t);
     }
     else
     {
         t = *ww / piv;
-        dd = store * sqrt(1.0 + t * t);
+        dd = store * sqrtf(1.0 + t * t);
     }
     *cos_ = *ww / dd;
     *sin_ = piv / dd;
